Added outputs for the AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY
The new subcommand 'id' prints the AWS_ACCESS_KEY_ID of the currently
used profile. The also added command 'key' prints the regarding
AWS_SECRET_ACCESS_KEY. For an easier export of these ID and KEY, the
command 'export' returns the export commands. This command can be used
for example together with the eval command to create an oneliner
exporting the credentials to the current shell.